Fixed various compiler warnings

+++ src/entities/entity102.cpp
@@ -521,7 +521,7 @@ bool IGES_ENTITY_102::ReadPD( std::ifstream& aFile, int& aSequenceVar )
     char pd = parent->globalData.pdelim;
     char rd = parent->globalData.rdelim;
 
-    idx = pdout.find( pd );
+    idx = (int)pdout.find( pd );
 
     if( idx < 1 || idx > 8 )
     {
