+++ src/entities/entity142.cpp
@@ -390,7 +390,7 @@ bool IGES_ENTITY_142::ReadPD( std::ifstream& aFile, int& aSequenceVar )
     char pd = parent->globalData.pdelim;
     char rd = parent->globalData.rdelim;
 
-    idx = pdout.find( pd );
+    idx = (int)pdout.find( pd );
 
     if( idx < 1 || idx > 8 )
     {
