+++ src/entities/entity144.cpp
@@ -300,7 +300,7 @@ bool IGES_ENTITY_144::format( int &index )
         return false;
     }
 
-    N2 = PTI.size();
+	N2 = (int)PTI.size();
 
     iPTS = PTS->GetDESequence();
     iPTO = PTO->GetDESequence();
@@ -512,7 +512,7 @@ bool IGES_ENTITY_144::ReadPD( std::ifstream& aFile, int& aSequenceVar )
     char pd = parent->globalData.pdelim;
     char rd = parent->globalData.rdelim;
 
-    idx = pdout.find( pd );
+	idx = (int)pdout.find(pd);
 
     if( idx < 1 || idx > 8 )
     {
